elo/requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
django-tastypie==0.14.0
idna==2.6
oauthlib==2.0.6
pkg-resources==0.0.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
style==1.1.0
update==0.0.1
urllib3==1.22
